+++ src/utils/clack-utils.ts
@@ -295,11 +295,13 @@ export async function confirmContinueIfPackageVersionNotSupported({
   packageName,
   packageVersion,
   acceptableVersions,
+  note,
 }: {
   packageId: string;
   packageName: string;
   packageVersion: string;
   acceptableVersions: string;
+  note?: string;
 }): Promise<void> {
   return traceStep(`check-package-version`, async () => {
     Sentry.setTag(`${packageName.toLowerCase()}-version`, packageVersion);
@@ -321,10 +323,8 @@ export async function confirmContinueIfPackageVersionNotSupported({
     );
 
     clack.note(
-      `Please upgrade to ${acceptableVersions} if you wish to use the Sentry Wizard.
-Or setup using ${chalk.cyan(
-        'https://docs.sentry.io/platforms/react-native/manual-setup/manual-setup/',
-      )}`,
+      note ??
+        `Please upgrade to ${acceptableVersions} if you wish to use the Sentry Wizard.`,
     );
     const continueWithUnsupportedVersion = await abortIfCancelled(
       clack.confirm({